이미지 수정 완료, 병원 리스트 검색하기 완료

+++ src/main/java/com/pawin/post/bo/PostImageBO.java
@@ -60,15 +60,18 @@ public class PostImageBO {
 			// 파일메니저 파일 갯수만큼 반복
 			for (MultipartFile file : files) {
 				// 파일이 있을 때만 업로드 -> 이미지 경로를 얻어냄
+				imagePaths = fileManagerService.saveFile(loginId, file);
 				for (ImagePath imagePath : imagePathList) {
 					// 이미지 패스도 null, 기존 이미지패스도 null이 아니면
-					//if (imagePaths != null && imagePath.getImagePath() != null) {
+					if (imagePaths != null && imagePath.getImagePath() != null) {
 						// 이미지 제거
 						fileManagerService.deleteFile(imagePath.getImagePath());
-					//}
+						postImageDAO.deleteImage(userId, postId);
+
+					}
 				}
-				imagePaths = fileManagerService.saveFile(loginId, file);
-				postImageDAO.updateImage(imagePaths, userId, postId);
+				postImageDAO.insertImagePost(imagePaths, userId, postId);
+
 			}
 
 		}
@@ -106,7 +109,7 @@ public class PostImageBO {
 
 			User user = userBO.getUserById(imagepath.getUserId());
 			imagePathView.setUser(user);
-			
+
 			imagepath.setPostId(postId);
 
 			// 결과물에 넣기
